messaging and group part completed

+++ src/components/Group/GroupList.js
@@ -11,21 +11,44 @@ import ModalWindow from "../../utils/ModalWindow";
 import GroupAddModal from "./GroupAddModal";
 import Swal from "sweetalert2";
 import Switch from "react-switch";
+import { jwtDecode } from "jwt-decode";
 
 const getAuthToken = () => localStorage.getItem("authToken");
+
+const getUserRoleFromToken = () => {
+  const token = getAuthToken();
+  if (!token) return null;
+
+  try {
+    const decodedToken = jwtDecode(token);
+    const role = decodedToken.role;
+    console.log(role);
+    return role;
+  } catch (error) {
+    console.error("Error decoding token:", error);
+    return null;
+  }
+};
 
 const GroupList = () => {
   const [groupList, setGroupList] = useState([]);
   const [addGroupModal, setAddGroupModal] = useState(false);
   const [editData, setEditData] = useState(null);
   const [edit, setEdit] = useState(false);
+  const [userRole, setUserRole] = useState(null);
 
   useEffect(() => {
-    fetchGroupList();
+    const role = getUserRoleFromToken();
+    setUserRole(role);
+    if (role !== "student") {
+      fetchGroupList();
+    }
   }, []);
 
   const fetchGroupList = async () => {
     const token = getAuthToken();
+    if (!token) return;
+
     try {
       const response = await axios.get(
         `${process.env.REACT_APP_API_BASE_URL}/group/groups`,
@@ -37,7 +60,6 @@ const GroupList = () => {
       );
       setGroupList(response.data.message);
     } catch (error) {
-      console.error("Error fetching groups:", error);
       Swal.fire({
         icon: "error",
         title: "Error",
@@ -52,6 +74,8 @@ const GroupList = () => {
 
   const addGroup = async (groupData) => {
     const token = getAuthToken();
+    if (!token || userRole === "student") return;
+
     try {
       const response = await axios.post(
         `${process.env.REACT_APP_API_BASE_URL}/group/group`,
@@ -68,7 +92,6 @@ const GroupList = () => {
         toggleAddGroupModal(); // Close the modal after adding
       }
     } catch (error) {
-      console.error("Error adding group:", error);
       Swal.fire({
         icon: "error",
         title: "Error",
@@ -78,6 +101,7 @@ const GroupList = () => {
   };
 
   const editGroup = (group) => {
+    if (userRole === "student") return;
     setEdit(true);
     setEditData(group);
     toggleAddGroupModal(); // Open the modal for editing
@@ -85,6 +109,8 @@ const GroupList = () => {
 
   const handleEditGroup = async (groupData) => {
     const token = getAuthToken();
+    if (!token || userRole === "student") return;
+
     try {
       const response = await axios.patch(
         `${process.env.REACT_APP_API_BASE_URL}/group/groups/${editData._id}`,
@@ -95,30 +121,20 @@ const GroupList = () => {
           },
         }
       );
-      console.log("Full response:", response);
-      if (response && response.status === 200) {
+
+      if (response.status === 200) {
         const updatedGroupList = groupList.map((group) =>
           group._id === editData._id ? { ...response.data.message } : group
         );
         setGroupList(updatedGroupList);
-
         toggleAddGroupModal();
-
         Swal.fire({
           icon: "success",
           title: "Group Updated",
           text: "The group has been updated successfully!",
         });
-      } else {
-        console.error("Unexpected response or status:", response);
-        Swal.fire({
-          icon: "error",
-          title: "Error",
-          text: "Unexpected response from server.",
-        });
       }
     } catch (error) {
-      console.error("Error updating group:", error);
       Swal.fire({
         icon: "error",
         text: "An error occurred while updating the group.",
@@ -128,6 +144,8 @@ const GroupList = () => {
 
   const deleteGroup = async (groupId) => {
     const token = getAuthToken();
+    if (!token || userRole === "student") return;
+
     try {
       const response = await axios.delete(
         `${process.env.REACT_APP_API_BASE_URL}/group/groups/${groupId}`,
@@ -140,7 +158,6 @@ const GroupList = () => {
 
       if (response.status === 200) {
         setGroupList(groupList.filter((group) => group._id !== groupId));
-
         Swal.fire({
           icon: "success",
           title: "Group Deleted",
@@ -148,7 +165,6 @@ const GroupList = () => {
         });
       }
     } catch (error) {
-      console.error("Error deleting group:", error);
       Swal.fire({
         icon: "error",
         title: "Error",
@@ -158,6 +174,8 @@ const GroupList = () => {
   };
 
   const toggleAtRiskStatus = async (groupId, currentAtRisk) => {
+    if (userRole === "student") return;
+
     const newAtRiskStatus = !currentAtRisk; // Toggle the current status
     const token = getAuthToken();
 
@@ -173,22 +191,10 @@ const GroupList = () => {
       );
 
       if (response.status === 200) {
-        // Update the group list
         const updatedGroupList = groupList.map((group) =>
           group._id === groupId ? { ...group, atRisk: newAtRiskStatus } : group
         );
         setGroupList(updatedGroupList);
-
-        // Add a notification for the at-risk change
-        addNotification({
-          type: "at-risk",
-          message: `Group "${
-            groupList.find((g) => g._id === groupId).name
-          }" has been flagged as ${
-            newAtRiskStatus ? "at risk" : "not at risk"
-          }.`,
-          date: new Date(),
-        });
 
         Swal.fire({
           icon: "success",
@@ -199,29 +205,21 @@ const GroupList = () => {
     } catch (error) {
       Swal.fire({
         icon: "error",
-        title: "Error",
         text: "An error occurred while toggling the 'at risk' status.",
       });
     }
   };
 
-  const addNotification = (notification) => {
-    const storedNotifications = localStorage.getItem("notifications");
-    const existingNotifications = storedNotifications
-      ? JSON.parse(storedNotifications)
-      : [];
-    const updatedNotifications = [...existingNotifications, notification];
-    localStorage.setItem("notifications", JSON.stringify(updatedNotifications));
-  };
-
   return (
     <div className="dataContainerBox">
-      <AppButton
-        name="Add Group"
-        customStyle="addBtnColor"
-        icon={faPlus}
-        onClick={toggleAddGroupModal}
-      />
+      {userRole !== "student" && ( // Only show if the user is not a student
+        <AppButton
+          name="Add Group"
+          customStyle="addBtnColor"
+          icon={faPlus}
+          onClick={toggleAddGroupModal}
+        />
+      )}
       <table className="table customTable mt-3">
         <thead className="text-center">
           <tr>
@@ -235,7 +233,7 @@ const GroupList = () => {
         <tbody>
           {groupList.length > 0 ? (
             groupList.map((group) => (
-              <tr key={group._id} className="text-center ">
+              <tr key={group._id} className="text-center">
                 <td className="tableData">{group.name}</td>
                 <td>{group.instructor ? group.instructor.username : "N/A"}</td>
                 <td>
@@ -252,22 +250,37 @@ const GroupList = () => {
                       }
                       checked={group.atRisk}
                       boxShadow="0px 1px 5px rgba(0, 0, 0, 0.5)"
-                      offColor="#2DBFCD" // Aesthetic choice for 'off' state
-                      onColor="#FFA500" // Aesthetic choice for 'on' state
-                      uncheckedIcon={false} // Hides the icons for simplicity
+                      offColor="#2DBFCD" // 'Off' color
+                      onColor="#FFA500" // 'On' color
+                      uncheckedIcon={false}
                       checkedIcon={false}
                       height={18}
                       width={36}
+                      // style={{
+                      //   cursor:
+                      //     userRole === "student" ? "not-allowed" : "pointer",
+                      // }}
+                      disabled={userRole === "student"} // Disable for students
                     />
                     <FontAwesomeIcon
                       icon={faPenToSquare}
                       className="actionIcons editIcon"
                       onClick={() => editGroup(group)}
+                      style={{
+                        cursor:
+                          userRole === "student" ? "not-allowed" : "pointer",
+                      }}
+                      // disabled={userRole === "student"} // Disable for students
                     />
                     <FontAwesomeIcon
                       icon={faTrashAlt}
                       className="actionIcons deleteIcon"
                       onClick={() => deleteGroup(group._id)}
+                      style={{
+                        cursor:
+                          userRole === "student" ? "not-allowed" : "pointer",
+                      }}
+                      disabled={userRole === "student"} // Disable for students
                     />
                   </div>
                 </td>
@@ -285,16 +298,16 @@ const GroupList = () => {
 
       <ModalWindow
         modal={addGroupModal}
-        toggleModal={toggleAddGroupModal} // This is correct
+        toggleModal={toggleAddGroupModal}
         modalHeader={edit ? "Edit Group" : "Add Group"}
         size="lg"
         modalBody={
           <GroupAddModal
             edit={edit}
             editData={editData}
-            toggleModal={toggleAddGroupModal} // Make sure this is passed
-            onSubmit={edit ? handleEditGroup : addGroup} // Ensure correct handling
-            onGroupAdded={fetchGroupList} // If there's a prop for after a group is added
+            toggleModal={toggleAddGroupModal}
+            onSubmit={edit ? handleEditGroup : addGroup}
+            disabled={userRole === "student"} // Disable for students
           />
         }
       />
